+++ dataBaseCreator.py
@@ -8,10 +8,7 @@ imgtitle = [(text)]
 #connection.commit()
 
 
-max_id_db = db.execute("SELECT MAX(id) FROM org")
-max_id = max_id_db.fetchone()[0]
-imgcount = 0
-if max_id != None:
-    imgcount = max_id
 
-print(imgcount)
+org_signal_list_db = db.execute("SELECT id, name FROM org")
+org_signal_list =org_signal_list_db.fetchall()
+print(org_signal_list[0][0])